service/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isodate==0.7.2
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
langchain==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.143
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
more-itertools==10.5.0
multidict==6.1.0
numpy==1.26.4
openai==1.54.4
openapi-core==0.19.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
orjson==3.10.11
packaging==24.2
parse==1.20.2
pathable==0.4.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.21.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
yarl==1.17.2
